forum/requirements.txt
# Python >= 3.8 required
Django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.14
python-dotenv>=1.0
pylint==3.2.7
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
django-phonenumber-field==8.0.0
phonenumberslite==9.0.0
djangorestframework-simplejwt==5.3.1
channels==4.0.0
daphne==4.0.0
mongoengine>=0.24.0
dnspython